feat: 🎸 create main methods

+++ src/constants/defines.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 use tokio;
 
-pub const DIM: u32 = 15;
+pub const DIM: usize = 15;
 pub const JSON_FILE_NAME: &str = "tile_simple.json";
 pub const GAME_WIDTH: u32 = 600;
 pub const GAME_HEIGHT: u32 = 600;
